depth to float

+++ includes/fdf.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   fdf.h                                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mbraets <mbraets@student.42.fr>            +#+  +:+       +#+        */
+/*   By: hoppy <hoppy@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/09 15:38:20 by mbraets           #+#    #+#             */
-/*   Updated: 2022/03/02 18:40:01 by mbraets          ###   ########.fr       */
+/*   Updated: 2022/03/02 22:19:47 by hoppy            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,7 +44,7 @@ typedef struct s_fdf {
 	float		angle;
 	int			zoom;
 	int			**map;
-	int			depth;
+	float		depth;
 	// int			height; //y
 	// int			width; //x
 	t_vector2	scr_size;
